+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unpubished
+
+### Added
+
+- Sigmoid calibration additional to isotonic
+
+### Changed
+
+- Moved Binarizer to different module
+- Moved PSL hyperparameters to constructor
+
 ## 0.4.2 - 2023-11-09
 
 ### Fixed